+++ app/Filament/Pages/VisitorStatistics.php
@@ -5,6 +5,7 @@ namespace App\Filament\Pages;
 use Filament\Pages\Page;
 use App\Filament\Widgets\VisitorStatsOverviewWidget;
 use App\Filament\Widgets\VisitorChartWidget;
+use App\Filament\Widgets\VisitorDeviceChartWidget;
 use App\Filament\Widgets\TopPagesTableWidget;
 use App\Filament\Widgets\TopLocationsTableWidget;
 
@@ -23,6 +24,7 @@ class VisitorStatistics extends Page
         return [
             VisitorStatsOverviewWidget::class,
             VisitorChartWidget::class,
+            VisitorDeviceChartWidget::class,
             TopLocationsTableWidget::class,
             TopPagesTableWidget::class,
         ];